Disable application termination from outside the main Application class
If methods are used as a library in another application, the application has
to handle the thrown errors itself. Thus we cannot simply call System.exit().

Subpackages of this program should simply propagate unrecoverable Exceptions,
the Application class main loop will terminate if it catches any Exception.
In this case, finalizers are run before the Application exits with system code 1.