Refactor "datacontainers"  package
The new classes are intended to be more flexible than the old ones.
Additionally, the functionality is now defined by interfaces, so that the actual implementations
can be echanged easily if required.

This commit introduces the following changes:
* Remove old PointListList related classes and introduce own data reprenentation interfaced
by PointList and PointListContainer interfaces.
* Reuse Point classes from the "printabledata package (which are now in their own java package).
* Adjust classes in "diagrams" package to the new data container.
(only Barchart yet, other classes need to be adjusted later on)
* Adjust "csvparser" classes to the new data container.
* Apply small changes to the current BarChartRenderer in order to make things run.
* Add example hardcoded renderer call in App.run() and log the rendered MatrixData.