Introduce new CSVParser
In contrast to the old implementation used in "SVGPlott", this approach uses
a CSVToBean reader.
This way, we do not have to care about the parsing process itself but can instead
define the elements of a CSV element (each element is basically one line in the file)
by a plain old java object, using only simple annotations.

The problem that arises is the current definition of the CSV files, which deviates from
regular csv data definitions.
This is why currently this approach does not manage to parse the first line of the custom CSV files,
this will need to be taken care of in a later commit.