Add line chart CSV parser
Currently, there arise multiple problems using the Bean parser.
The current implementation does seem to evaluate the number of columns on the first row only,
and thus ignores elements on the second line, if they extend the number of elements on the first row.
Additionally, parsing doubles with leading whitespaces does seem to be a problem.

This is why I decided to end development on this branch, and continue the work on the existing
parsers fromCSVPlott.
Later on, one might want to continue this branch.