Add multi page support
For each data set in the input data, a new page is created in the rendering process.
The scale factor is shared across all diagramms and  is determined by analyzing
the maximum value of all datasets.
Following this, all axis and their tickmarks are shared across all diagrams.

This ensures comparability between the points on different pages.